Disable chunking optimization in CRA
First thing I noticed is that the CRA react-scripts version was at 3.1.1
which is the latest and these ejecting libraries are waaaay behind.

So first thing I did was downgrade the react-scripts version.

Note: both craco and react-app-rewired work.

Then the next thing was I used the [LimitChunkCount
Plugin](https://webpack.js.org/plugins/limit-chunk-count-plugin/) to set
the maximum number of the chunks to 1. This seemed to work.